Revert "Add time versus value plot and histogram"
