+++ app/build.gradle
@@ -54,6 +54,4 @@ dependencies {
     implementation 'com.google.firebase:firebase-analytics'
     implementation 'com.google.firebase:firebase-firestore'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
-    implementation 'com.jjoe64:graphview:4.2.2'
-
 }
