+++ app/src/main/AndroidManifest.xml
@@ -9,21 +9,9 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Theme.PreemptiveOOP">
-        <activity android:name=".experiment.DisplayExpHistogram">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+        <activity android:name=".experiment.DisplayExpStats"></activity>
+        <activity android:name=".experiment.DisplayExpGraphs" ></activity>
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity android:name=".experiment.DisplayExpStats" />
-        <activity android:name=".experiment.DisplayExpPlots">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
         <activity android:name=".trial.ExecuteTrial" />
         <activity android:name=".experiment.ExperimentList" />
         <activity android:name=".user.UserRegister" />
