+++ app/src/main/res/layout/activity_display_exp_graphs.xml
@@ -6,19 +6,11 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="100dp"
-        android:text="  Histogram"
-        android:textSize="40sp"
-        android:textStyle="bold" />
-
 
     <com.github.mikephil.charting.charts.BarChart
-        android:layout_marginTop="200dp"
-        android:id="@+id/gv_histogram"
+        android:id="@+id/graphicv_exp_display_stats"
         android:layout_width="match_parent"
-        android:layout_height="400dp" />
+        android:layout_height="match_parent" />
 
 
 </LinearLayout>
