--- app/src/main/res/layout/activity_display_exp_plot.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/tv_plot_description"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:text=" Time vs Value  Plot"
-        android:textSize="40sp"
-        android:textStyle="bold" />
-
-
-    <com.jjoe64.graphview.GraphView
-        android:id="@+id/gv_plot"
-        android:layout_marginLeft="30dp"
-        android:layout_marginRight="20dp"
-        android:layout_width="350dp"
-        android:layout_height="300dp"
-        android:layout_marginTop="20dp" />
-
-</LinearLayout>
